Add IPAPPEND directive to PXE boot menu entries
Adding the `IPAPPEND 2` directive to the PXE boot menu entries will
cause pxelinux to add `BOOTIF=01-{mac address of dhcp interface}` to the
kernel command line.  The initramfs scripts use this setting to identify
which network interface to use for the remainder of the netboot process.
This allows the system to boot properly regardless of the names assigned
to the network interfaces.