Do NOT use net.ifnames=0 on kernel comamnd line
This lets the kernel use what it wants to use for network interface
names.  Users wishing more control can use udev rules or more modern
systemd related ways of renaing interfaces.  While not using the modern
kernel naming convention results in the more "traditional" `ethX` names,
it also makes renaming interfaces to a preferred-but-different `ethY`
ordering much more complicated and prone to problems due to the
potential for desired names to be already claimed by (or assigned to)
other interfaces.