Make /persistent/etc/udev/rules.d with tmpfiles.d
This is part of netboot-base.conf so it will be done at bootstrap time.
This adds /persistent/etc/udev/rules.d to netboot-base.conf such that it
will get created if it doesn't exist at boot time and a symlink to it
from /etc/udev/rules.d will be created.  This will run as part of the
bootstrap process so the link will be created from a chroot on the head
node.  The directory will not be created automatically until/unless the
host boots, so playbooks that want to put files in
/persistent/etc/udev/rules.d should ensure this directory exists (by
creating it if necessary) before placing files in it.