Include from another directory in auto.master
auto.master map files can now be dropped in
/persistent/etc/auto.master.d in addition to /etc/auto.master.d.  Files
in the former will take precedence over files of the same name in the
latter.  One possible use of this is to put `ro` (read-only) options in
a master map file in /etc/auto.master.d.  Then for hosts that require
`rw` (read-write) access, the file can be copied to
/persistent/etc/auto.master.d with the `ro` option changed to `rw`.